refactor: Mv testFromStyleguidist out of src directory
